+++ app/sprinkles/core/src/Core.php
@@ -10,7 +10,6 @@
 
 namespace UserFrosting\Sprinkle\Core;
 
-use Psr\Container\ContainerInterface;
 use RocketTheme\Toolbox\Event\Event;
 use UserFrosting\Sprinkle\Core\Csrf\SlimCsrfProvider;
 use UserFrosting\Sprinkle\Core\Database\Models\Model;
@@ -36,18 +35,6 @@ class Core extends Sprinkle
     ];
 
     /**
-     * Create a new Sprinkle object.
-     *
-     * @param ContainerInterface $ci The global container object, which holds all your services.
-     */
-    public function __construct(ContainerInterface $ci)
-    {
-        $this->ci = $ci;
-
-        $this->registerStreams();
-    }
-
-    /**
      * Defines which events in the UF lifecycle our Sprinkle should hook into.
      */
     public static function getSubscribedEvents()
@@ -70,6 +57,8 @@ class Core extends Sprinkle
 
         // Set container for environment info class
         EnvironmentInfo::$ci = $this->ci;
+
+        $this->registerStreams();
     }
 
     /**