+++ app/sprinkles/core/src/Twig/CacheHelper.php
@@ -22,6 +22,8 @@ class CacheHelper
 {
     /**
      * @var ContainerInterface The global container object, which holds all your services.
+     *
+     * @todo Change this, only the locator service is required
      */
     protected $ci;
 
@@ -38,7 +40,7 @@ class CacheHelper
     /**
      * Function that delete the Twig cache directory content.
      *
-     * @return bool true/false if operation is successfull
+     * @return bool true/false if operation is successful
      */
     public function clearCache()
     {